dto2

+++ WeatherApp/app/src/main/java/com/example/weatherapplication/data/networking/models/forecast/response/nesteds/Wind.kt
@@ -8,6 +8,6 @@ data class Wind(
     val speed: Double,
 
     @SerializedName("deg")
-    val routeDegrees: Int
+    val routeDegrees: Float
 )
